Fix export, ft_replace, added ft_strcmp

+++ src/interpreter.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   interpreter.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
+/*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/07 00:16:07 by jvacaris          #+#    #+#             */
-/*   Updated: 2021/11/11 01:04:41 by jvacaris         ###   ########.fr       */
+/*   Updated: 2021/11/14 20:16:38 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -111,15 +111,15 @@ static void	command_finder(char **list, int *idx, t_str **env_list)
 	assist = adv_qm_rem(list[*idx], 0);
 	if (!ft_strncmp(assist, "echo\0", 5))
 		ft_echo(&list[*idx]);
-	else if (!ft_strncmp(assist, "pwd", INT_MAX))
+	else if (!ft_strncmp(assist, "pwd", ULONG_MAX))
 		ft_pwd(env_list, &list[*idx]);
-	else if (!ft_strncmp(assist, "env", INT_MAX))
+	else if (!ft_strncmp(assist, "env", ULONG_MAX))
 		ft_env(env_list, &list[*idx]);
-	else if (!ft_strncmp(assist, "unset", INT_MAX))
+	else if (!ft_strncmp(assist, "unset", ULONG_MAX))
 		ft_unset(env_list, &list[*idx]);
-	else if (!ft_strncmp(assist, "export", INT_MAX))
+	else if (!ft_strncmp(assist, "export", ULONG_MAX))
 		ft_export(env_list, &list[*idx]);
-	else if (!ft_strncmp(assist, "cd", INT_MAX))
+	else if (!ft_strncmp(assist, "cd", ULONG_MAX))
 		ft_cd(env_list, &list[*idx]);
 	else
 		exec_command(&list[*idx], env_list_to_vector(env_list));
